updated constructor to take float instead of int D:

+++ Project3/node.cpp
@@ -14,7 +14,7 @@ Node::Node(int d)
 	next = NULL;
 }
 
-Node::Node(int d, int w)
+Node::Node(int d, float w)
 {
 	num = d;
 	weight = w;
